move programs to separate folder and cleanup
