+++ README.md
@@ -31,6 +31,15 @@ You're all set up baby!
 python color_change.py
 ```
 
+## Light Programs
+A program dynamically updates the lights and generally continues running forever. 
+These all live in the `programs` directory.
+
+To run the main of a program without running the server, run the following from the root directory
+ to trigger the `if __name__ == "__main__"` section of a program:
+```bash
+python -m programs.color_cycle
+```
 ## Run the server
 Run `./start_server.sh` to start a localhost server
 
@@ -45,10 +54,10 @@ You should receive a nice message.
 This just sends a GET http request to the address specified.
 
 From a high level, a get request just allows us to call a function on a server
-without any arguments. So the following is like calling `run_color_sequence()` in a python script:
+without any arguments. So the following is like calling `run_color_cycle()` in a python script:
 
 ```
-curl http://127.0.0.1:8000/run_color_sequence/
+curl http://127.0.0.1:8000/run_color_cycle/
 ```
 
 The other api method currently supported is:
