--- halloween_arduino/bathroom_lights/bulb_info_2_dict.json
@@ -1,14 +1,0 @@
-{
-    "192.168.68.79": {
-        "room": "bathroom",
-        "name": "bathroom_1"
-    },
-    "192.168.68.78": {
-        "room": "bathroom",
-        "name": "bathroom_2"
-    },
-    "192.168.68.77": {
-        "room": "bathroom",
-        "name": "bathroom_3"
-    }
-}
