--- halloween_arduino/bathroom_lights/bulb_info_3_dict.json
@@ -1,6 +1,0 @@
-{
-    "192.168.68.80": {
-        "room": "on_amp",
-        "name": "living_room"
-    }
-}
