--- halloween_arduino/halloween_min_viable_product/Arduino_button/.vscode/extensions.json
@@ -1,7 +1,0 @@
-{
-    // See http://go.microsoft.com/fwlink/?LinkId=827846
-    // for the documentation about the extensions.json format
-    "recommendations": [
-        "platformio.platformio-ide"
-    ]
-}
