--- halloween_arduino/halloween_min_viable_product/bathroom_lights/bulb_info.json
@@ -1,9 +1,0 @@
-{"192.168.68.61": {"room": "living room", "name": "box light"}, 
-"192.168.68.60": {"room": "living room", "name": "pole side light"}, 
-"192.168.68.63": {"room": "office", "name": "shay's desk light"}, 
-"192.168.68.58": {"room": "living room", "name": "big lamp"}, 
-"192.168.68.59": {"room": "bed room", "name": "shay's light"}, 
-"192.168.68.57": {"room": "living room", "name": "pole top light"}, 
-"192.168.68.56": {"room": "office", "name": "spot light"}, 
-"192.168.68.62": {"room": "office", "name": " madie's desk lamp"}, 
-"192.168.68.55": {"room": "bed room", "name": "madie's light"}}
