--- halloween_arduino/halloween_min_viable_product/bathroom_lights/bulb_info_2_dict.json
@@ -1,14 +1,0 @@
-{
-    "192.168.68.58": {
-        "room": "bathroom",
-        "name": "bathroom_1"
-    },
-    "192.168.68.55": {
-        "room": "bathroom",
-        "name": "bathroom_2"
-    },
-    "192.168.68.56": {
-        "room": "bathroom",
-        "name": "bathroom_3"
-    }
-}
