--- halloween_arduino/halloween_min_viable_product/bathroom_lights/wiz_lights.code-workspace
@@ -1,8 +1,0 @@
-{
-	"folders": [
-		{
-			"path": "."
-		}
-	],
-	"settings": {}
-}
